ANDROID: scsi: ufs: set crypto keyslot before prepare_command
Move the call to ufshcd_prepare_lrbp_crypto() to before
trace_android_vh_ufs_prepare_command() rather than after.

This allows trace_android_vh_ufs_prepare_command() to adjust the crypto
keyslot number in order to support nonstandard hardware that numbers the
crypto keyslots as 1..N instead of 0..N-1 as expected.

Alternatively we could add a hook just for remapping the keyslot number.
But it seems better to just use the existing prepare_command hook.

This should be folded into
ANDROID-scsi-ufs-add-a-hook-to-prepare-command.patch.

Bug: 166139333
Bug: 162257402
Change-Id: I5f4e3a0d3322145ded4e70e011446fb51825877e
Signed-off-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: Matthias Maennich <maennich@google.com>