USB: PHY: Fix PHY Power Sequence for Shared LDOs in USB HS PHY
This commit addresses an issue in the PHY power on/off sequence for USB
controllers using HS PHY, where power rails are shared across multiple
controllers. The issue arises when the recommended power sequence from
SNPS is violated due to the use of shared LDOs.

The recommended power-on sequence is:
1. Assert reset from all PHYs (BCR=1)
2. Power On LDOs (VDDA, 1.8 then 3.1)
3. Corresponding controller in collapse (GDSC, SW Collapse=0)
4. De-Assert reset from all PHYs (BCR=0)

The recommended power-off sequence is:
1. Assert reset from all PHYs (BCR=1)
2. Corresponding controller in collapse (GDSC, SW Collapse=1)
3. Power off LDOs (VDDA, 3.1 then 1.8)
4. De-Assert reset from all PHYs (BCR=0)

In scenarios where LDOs are shared between multiple USB controllers,
there is a chance of violating this sequence during power on and off.

For example, consider the following scenario:

USB0 HS-PHY:  LDO-A (vdd18) and LDO-B(vdd31)
USB1 HS-PHY:  LDO-C (vdd18) and LDO-B(vdd31)

Here LDO-B(vdd31) is shared between USB0-HS and USB1-HS PHYs. In this
scenario, there is a chance of violating the sequence in power on and
off.

This commit fixes this issue by treating non-shared LDOs (LDO-C(vdd18)
and LDO-A(vdd18)) as shared LDOs from a software perspective and
voting from software to maintain the correct sequence in both power on
and off scenarios.

Change-Id: Ibbb16c55e79b1ffe030005a30f07b98a700b0ecd
Signed-off-by: Selvarasu Ganesan <quic_selvaras@quicinc.com>